+++ crates/common/src/db_migrations/remove_unused_common_tables.rs
