+++ crates/common/src/lib.rs
@@ -4,8 +4,8 @@
 pub mod cache;
 pub mod config;
 pub mod da;
+pub mod db_migrations;
 pub mod error;
-pub mod migration;
 pub mod rpc;
 pub mod tasks;
 pub mod utils;
