+++ crates/fullnode/src/db_migrations/mod.rs
@@ -3,7 +3,7 @@ use std::sync::OnceLock;
 use sov_db::{ledger_db::migrations::LedgerMigration, schema::tables::FULL_NODE_LEDGER_TABLES};
 
 use crate::db_migrations::verified_batch_proof_encoding::FixVerifiedBatchProofsEncoding;
-use citrea_common::migration::{
+use citrea_common::db_migrations::{
     MigrateBatchAndSlotByNumber, MigrateVerifiedProofsBySlotNumber, RemoveUnusedTables,
 };
